route: move to locale
