+++ app/[locale]/(interactive)/4-2/page.tsx
