+++ app/[locale]/(non-interactive)/layout.tsx
